Port Library to JsonSchema
The unsupported nature of this app made it difficult to get going inside
of one of our kiosk implementations. The following changes make it
"work".

The problem in this library was getting the rather-unsupported JSV
library replaced with something a bit more modern for validating our
JSON schemas. Since the other app work uses jsonschema, I adopted the
same there.

Changes:
 - [x] Correct Imports to use JSON Schema
 - [x] Instantiate new validator
 - [x] Bump Version
 - [x] Correct package.json

Hey, it passes the spec suite after this change.